`Manual`: move around some code related to generated names
