+++ manual/c/generated-names/unprefixed_field_names.c
@@ -1,0 +1,1 @@
+#include "unprefixed_field_names.h"
