+++ manual/c/manual_examples.h
@@ -63,23 +63,6 @@ typedef struct {
 } *config;
 
 /**
- * Awkward names
- */
-
-typedef int adiós;
-// Required since Apple's GCC Assembler and LLVM IR does not allow Unicode characters
-#if defined(SUPPORTS_UNICODE)
-typedef int 数字;
-void 拜拜(void);
-void ϒ(void);
-#else
-void ByeBye(void);
-void Gamma(void);
-#endif
-typedef int data;
-void import(void);
-
-/**
  * Flavours of enums
  */
