--- manual/c/omit_field_prefixes.c
@@ -1,1 +1,0 @@
-#include "omit_field_prefixes.h"
